split sbml and pysb tests in two files
