+++ .github/workflows/build.yml
@@ -35,7 +35,7 @@ jobs:
       if: matrix.os == 'macos-latest'
       shell: bash -l {0}
       run: |
-        pytest pydyno --deselect=item=pydyno/tests/test_discretization.py::TestPathSbmlSingle
+        pytest pydyno --ignore=pydyno/tests/test_discretization_sbml.py
     - name: Test with pytest in ubuntu and windows
       if: matrix.os == 'ubuntu-latest' || matrix.os == 'windows-latest'
       shell: bash -l {0}
